Review and maintenance before public release, final
* Going full `size_t`.
  For simplicity, all "size" or "count" arguments are `size_t` now, except when negative values are allowed.
  Tbh, I quite like `int64_t` for everything and reserving unsigned integers for specific cases (e.g.
  bitmasks). However, the STL is using `size_t` everywhere and indexing is unsigned so if conversion warnings
  are activated it is just annoying to cast everything everytime. For now, choose `size_t`.

* CUDA: Switching to 2D blocks.
  Some kernels have been refactored to be fully batched and to use a more flexible 2D block. This should
  be similar or faster depending on the size of the innermost dimensions. I have no benchmarks though.

* Add `const` when I feel like it...

* Change from `cudaPeekAtLastError` to `cudaGetLastError`.
  When the error is retrieved and thrown, the current state should be reset to `cudaSuccess`.

* Simplifying API.
  For the backends, cases where 2D or 3D array differs for parameters other than the shape, use explicit
  functions like `sphere2D` and `sphere3D` as opposed to `sphere`.
  Also, remove overload for "output write" with `fft::lowpass`, `fft::highpass`, `fft::bandpass`,
  `filter::sphere*`, `filter::rectangle` and `filter::cylinder3D`. Instead, allow the input to `nullptr`.

* Fixes for the last few commits that introduced some errors.

* Move `fastSize` to `fft` namespace.

* Add `index()` and `coordinates()`.

Tags: [fix][maintenance][refactor]
Compile: yes
Tests: yes

+++ src/noa/common/transform/Symmetry.cpp
@@ -165,22 +165,22 @@ namespace {
     //@CLION-formatter:on
 
     // Axial on Z.
-    constexpr void setCX(float33_t* rotm, uint order) {
+    constexpr void setCX(float33_t* rotm, size_t order) {
         NOA_ASSERT(order > 0);
         const float angle = math::Constants<float>::PI2 / static_cast<float>(order);
-        for (uint i = 1; i < order; ++i) // skip the identity
+        for (size_t i = 1; i < order; ++i) // skip the identity
             rotm[i - 1] = transform::rotateZ(static_cast<float>(i) * angle);
     }
 
     // Axial on Z, plus 2-fold on X
-    constexpr void setDX(float33_t* rotm, uint order) {
+    constexpr void setDX(float33_t* rotm, size_t order) {
         NOA_ASSERT(order > 0);
         setCX(rotm, order);
         float33_t two_fold_x(+1, +1, +1,
                              -1, -1, +1,
                              +1, +1, -1);
         rotm[order - 1] = math::elementMultiply(float33_t(), two_fold_x); // -1 since the identity isn't there
-        for (uint i = 0; i < order - 1; ++i)
+        for (size_t i = 0; i < order - 1; ++i)
             rotm[order + i] = math::elementMultiply(rotm[i], two_fold_x);
     }
 }
