+++ src/noa/cpu/transform/Apply.cpp
@@ -13,10 +13,10 @@ namespace {
     // 2D, 2x3 matrices
     template<typename T, InterpMode INTERP, BorderMode BORDER>
     void apply_(const T* input, size2_t input_shape, T* outputs, size2_t output_shape,
-                const float23_t* transforms, uint nb_transforms, T value) {
+                const float23_t* transforms, size_t nb_transforms, T value) {
         cpu::transform::Interpolator2D<T> interp(input, input_shape, input_shape.x, value);
         float2_t coordinates;
-        for (uint i = 0; i < nb_transforms; ++i) {
+        for (size_t i = 0; i < nb_transforms; ++i) {
             for (size_t y = 0; y < output_shape.y; ++y) {
                 for (size_t x = 0; x < output_shape.x; ++x, ++outputs) {
                     coordinates = transforms[i] * float3_t(x, y, 1.f);
@@ -29,9 +29,9 @@ namespace {
     // 2D, 3x3 matrices
     template<typename T, InterpMode INTERP, BorderMode BORDER>
     void apply_(const T* input, size2_t input_shape, T* outputs, size2_t output_shape,
-                const float33_t* transforms, uint nb_transforms, T value) {
+                const float33_t* transforms, size_t nb_transforms, T value) {
         cpu::transform::Interpolator2D<T> interp(input, input_shape, input_shape.x, value);
-        for (uint i = 0; i < nb_transforms; ++i) {
+        for (size_t i = 0; i < nb_transforms; ++i) {
             float23_t transform(transforms[i]);
             for (size_t y = 0; y < output_shape.y; ++y) {
                 for (size_t x = 0; x < output_shape.x; ++x, ++outputs) {
@@ -47,10 +47,10 @@ namespace {
     // 3D, 3x4 matrices
     template<typename T, InterpMode INTERP, BorderMode BORDER>
     void apply_(const T* input, size3_t input_shape, T* outputs, size3_t output_shape,
-                const float34_t* transforms, uint nb_transforms, T value) {
+                const float34_t* transforms, size_t nb_transforms, T value) {
         cpu::transform::Interpolator3D<T> interp(input, input_shape, input_shape.x, value);
         float3_t coordinates;
-        for (uint i = 0; i < nb_transforms; ++i) {
+        for (size_t i = 0; i < nb_transforms; ++i) {
             for (size_t z = 0; z < output_shape.z; ++z) {
                 for (size_t y = 0; y < output_shape.y; ++y) {
                     for (size_t x = 0; x < output_shape.x; ++x, ++outputs) {
@@ -65,9 +65,9 @@ namespace {
     // 3D, 4x4 matrices
     template<typename T, InterpMode INTERP, BorderMode BORDER>
     void apply_(const T* input, size3_t input_shape, T* outputs, size3_t output_shape,
-                const float44_t* transforms, uint nb_transforms, T value) {
+                const float44_t* transforms, size_t nb_transforms, T value) {
         cpu::transform::Interpolator3D<T> interp(input, input_shape, input_shape.x, value);
-        for (uint i = 0; i < nb_transforms; ++i) {
+        for (size_t i = 0; i < nb_transforms; ++i) {
             float34_t transform(transforms[i]);
             for (size_t z = 0; z < output_shape.z; ++z) {
                 for (size_t y = 0; y < output_shape.y; ++y) {
@@ -85,7 +85,7 @@ namespace {
 
     template<typename T, typename SHAPE, typename MATRIX, InterpMode INTERP>
     void launch_(const T* input, SHAPE input_shape, T* outputs, SHAPE output_shape,
-                 const MATRIX* transforms, uint nb_transforms, T value, BorderMode border_mode) {
+                 const MATRIX* transforms, size_t nb_transforms, T value, BorderMode border_mode) {
         switch (border_mode) {
             case BORDER_ZERO:
                 apply_<T, INTERP, BORDER_ZERO>(
@@ -118,7 +118,7 @@ namespace {
 
     template<typename T, typename SHAPE, typename MATRIX>
     void launch_(const T* input, SHAPE input_shape, T* outputs, SHAPE output_shape,
-                 const MATRIX* transforms, uint nb_transforms, T value,
+                 const MATRIX* transforms, size_t nb_transforms, T value,
                  InterpMode interp_mode, BorderMode border_mode) {
         switch (interp_mode) {
             case INTERP_NEAREST:
@@ -150,7 +150,7 @@ namespace {
 namespace noa::cpu::transform {
     template<bool PREFILTER, typename T, typename MATRIX>
     void apply2D(const T* input, size2_t input_shape, T* outputs, size2_t output_shape,
-                 const MATRIX* transforms, uint nb_transforms,
+                 const MATRIX* transforms, size_t nb_transforms,
                  InterpMode interp_mode, BorderMode border_mode, T value) {
         NOA_PROFILE_FUNCTION();
         if (PREFILTER && interp_mode == INTERP_CUBIC_BSPLINE) {
@@ -166,7 +166,7 @@ namespace noa::cpu::transform {
 
     template<bool PREFILTER, typename T, typename MATRIX>
     void apply3D(const T* input, size3_t input_shape, T* outputs, size3_t output_shape,
-                 const MATRIX* transforms, uint nb_transforms,
+                 const MATRIX* transforms, size_t nb_transforms,
                  InterpMode interp_mode, BorderMode border_mode, T value) {
         NOA_PROFILE_FUNCTION();
         if (PREFILTER && interp_mode == INTERP_CUBIC_BSPLINE) {
@@ -180,15 +180,15 @@ namespace noa::cpu::transform {
         }
     }
 
-    #define NOA_INSTANTIATE_APPLY_(T)                                                                                               \
-    template void apply2D<true, T, float23_t>(const T*, size2_t, T*, size2_t, const float23_t*, uint, InterpMode, BorderMode, T);   \
-    template void apply2D<false, T, float23_t>(const T*, size2_t, T*, size2_t, const float23_t*, uint, InterpMode, BorderMode, T);  \
-    template void apply2D<true, T, float33_t>(const T*, size2_t, T*, size2_t, const float33_t*, uint, InterpMode, BorderMode, T);   \
-    template void apply2D<false, T, float33_t>(const T*, size2_t, T*, size2_t, const float33_t*, uint, InterpMode, BorderMode, T);  \
-    template void apply3D<true, T, float34_t>(const T*, size3_t, T*, size3_t, const float34_t*, uint, InterpMode, BorderMode, T);   \
-    template void apply3D<false, T, float34_t>(const T*, size3_t, T*, size3_t, const float34_t*, uint, InterpMode, BorderMode, T);  \
-    template void apply3D<true, T, float44_t>(const T*, size3_t, T*, size3_t, const float44_t*, uint, InterpMode, BorderMode, T);   \
-    template void apply3D<false, T, float44_t>(const T*, size3_t, T*, size3_t, const float44_t*, uint, InterpMode, BorderMode, T)
+    #define NOA_INSTANTIATE_APPLY_(T)                                                                                                 \
+    template void apply2D<true, T, float23_t>(const T*, size2_t, T*, size2_t, const float23_t*, size_t, InterpMode, BorderMode, T);   \
+    template void apply2D<false, T, float23_t>(const T*, size2_t, T*, size2_t, const float23_t*, size_t, InterpMode, BorderMode, T);  \
+    template void apply2D<true, T, float33_t>(const T*, size2_t, T*, size2_t, const float33_t*, size_t, InterpMode, BorderMode, T);   \
+    template void apply2D<false, T, float33_t>(const T*, size2_t, T*, size2_t, const float33_t*, size_t, InterpMode, BorderMode, T);  \
+    template void apply3D<true, T, float34_t>(const T*, size3_t, T*, size3_t, const float34_t*, size_t, InterpMode, BorderMode, T);   \
+    template void apply3D<false, T, float34_t>(const T*, size3_t, T*, size3_t, const float34_t*, size_t, InterpMode, BorderMode, T);  \
+    template void apply3D<true, T, float44_t>(const T*, size3_t, T*, size3_t, const float44_t*, size_t, InterpMode, BorderMode, T);   \
+    template void apply3D<false, T, float44_t>(const T*, size3_t, T*, size3_t, const float44_t*, size_t, InterpMode, BorderMode, T)
 
     NOA_INSTANTIATE_APPLY_(float);
     NOA_INSTANTIATE_APPLY_(double);
