+++ tests/noa/gpu/cuda/memory/TestCUDATranspose.cpp
@@ -79,7 +79,7 @@ TEMPLATE_TEST_CASE("cuda::memory::transpose() - random shapes - contiguous layou
 
     test::Randomizer<TestType> randomizer(-5., 5.);
     size3_t shape = test::getRandomShape(ndim);
-    uint batches = test::IntRandomizer<uint>(1, 4).get();
+    size_t batches = test::IntRandomizer<size_t>(1, 4).get();
     size_t elements = noa::elements(shape) * batches;
     cpu::memory::PtrHost<TestType> h_data(elements);
     test::initDataRandom(h_data.get(), elements, randomizer);
